modal separate from form CreateNewDocument
