+++ src/assets/scss/components/document/_Document.scss
@@ -159,6 +159,6 @@
 
   &__new-document-modal-footer{
     border-top: transparent;
-    padding: 0rem 1rem 1rem 1rem;
+    padding: 0rem 0rem 0rem 0rem;
   }
 }
