+++ src/assets/scss/styles.css
@@ -808,7 +808,7 @@ CSS for UserProfile
 
 .document__new-document-modal-footer {
   border-top: transparent;
-  padding: 0rem 1rem 1rem 1rem; }
+  padding: 0rem 0rem 0rem 0rem; }
 
 /*Clases do Home*/
 .o-thumb-circle {
