+++ src/components/modals/Modals.js
@@ -1,0 +1,7 @@
+import React, { Component } from 'react';
+import { connect } from 'react-redux';
+
+
+class Modals extends Component {
+
+}
