+++ src/helpers/config.js
@@ -1,3 +1,3 @@
-export const apiUrl = 'http://api.masteraula.com.br';
+export const apiUrl = 'https://api.masteraula.com.br';
 
 export default apiUrl;
